Add the github Oauth option to the dev login page
- Break out the local login into a new client side file to avoid
rehydration errors in nextjs v15
